+++ config/inflections.rb
@@ -15,6 +15,7 @@ HCloud.loader.inflector.inflect(
   "iso" => "ISO",
   "iso_type" => "ISOType",
   "primary_ip" => "PrimaryIP",
+  "primary_ip_prices" => "PrimaryIPPrices",
   "service_http" => "ServiceHTTP",
   "ssh_key" => "SSHKey",
   "target_ip" => "TargetIP",