+++ lib/hcloud/entities/primary_ip_prices.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module HCloud
+  class PrimaryIPPrices < Entity
+    attribute :type
+    attribute :prices, :price
+  end
+end